Fix uninitialized pointer dereference in media_device_add_entity
BZ:182374

When the entity being added has the default flag set, the function tries
to set the entity as the default entity for its type by storing the
entity pointer in the appropriate default entity field. If the entity
type isn't recognized, the default entity pointer is left uninitialized,
leading to a crash. Fix it.

Change-Id: I6791ac2a9ddcc43aaf6cf352a07c73c0f4ca7dab
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>